requirements.txt
pymupdf
reportlab
langchain
langchain-core
langchain-openai
langchain-text-splitters
langchain_postgres
python-decouple
django
frontend
tools
gunicorn
pdf2docx
psycopg
psycopg-binary
boto3